+++ tests/tests.py
@@ -1380,6 +1380,7 @@ def test_issue1281():
     run(dpath("test_issue1281"))
 
 
+@skip_on_windows  # TODO on windows, dot command is suddenly not found anymore although it is installed
 def test_filegraph():
     workdir = dpath("test_filegraph")
     dot_path = os.path.join(workdir, "fg.dot")
@@ -1391,13 +1392,11 @@ def test_filegraph():
         dot_path = dot_path.replace("\\", "/")
 
     # make sure the calls work
-    shell("cd {workdir};python -m snakemake --filegraph > {dot_path}")
+    shell("cd {workdir}; python -m snakemake --filegraph > {dot_path}")
 
     # make sure the output can be interpreted by dot
-    with open(dot_path, "rb") as dot_file, open(pdf_path, "wb") as pdf_file:
-        pdf_file.write(
-            subprocess.check_output(["dot", "-Tpdf"], stdin=dot_file, cwd=workdir)
-        )
+    shell("cd {workdir}; dot -Tpdf > {pdf_path} < {dot_path}")
+
     # make sure the generated pdf file is not empty
     assert os.stat(pdf_path).st_size > 0
 
@@ -1496,6 +1495,7 @@ def test_output_file_cache_remote():
 
 @connected
 @zenodo
+@pytest.mark.xfail(reason="zenodo currently returns an internal server error")
 def test_remote_zenodo():
     run(dpath("test_remote_zenodo"))
 
@@ -2058,3 +2058,8 @@ def test_localrule():
 @skip_on_windows
 def test_module_wildcard_constraints():
     run(dpath("test_module_wildcard_constraints"))
+
+
+@skip_on_windows
+def test_config_yte():
+    run(dpath("test_config_yte"))